Evaluate past events and datetimes to sort upcoming/past
Previously, the Calendar page was only moving Upcoming Events to Past and not the other way around. Also the method to do this was only respecting the datetimes of the initial event, not any recurrences that were further out. This change causes the method to look at every date related to the event and put it in Past or Upcoming appropriately.

Fixes https://trello.com/c/1emrN1Ra/88-events-that-are-not-past-are-showing-up-in-past-wtf