Allow filtering upcoming events by eventtype
This will allow you to filter events on the upcoming list to specific event types. For example:
`/upcoming?eventtype=camp` would only show events that have the 'camp' type
`/upcoming?eventtype=special%20event` would only show events that have the 'special event' type

This could then be used in conjunction with the framework events widget to only show upcoming events of a specific type. I don't think the type is being used anywhere else, so this might be a helpful way of using it.